[Loader] Pass around received buffers between threads
Currently WorkerLoaderClientBridgeSyncHelper retains all received data until
the loading finishes, but it is not necessary: All we need is that the data is
valid when the client (in a worker thread) receives it.

This change stops keeping received data and passes it to the worker thread.

BUG=None

Review URL: https://codereview.chromium.org/1648553002

Cr-Original-Commit-Position: refs/heads/master@{#372336}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ebfe8890f0f9bc994d5d5ea6698c704f1403b8f8