Add a new layout test case to canvas-ImageBitmap-transferable.html
In this CL, we add a new layout test case that applies structured clone
to a neutered ImageBitmap after it was transfered to a worker. The expected
behavior is to throw an exception to user notifying that the ImageBitmap
has been neutered.

Review URL: https://codereview.chromium.org/1652553003

Cr-Original-Commit-Position: refs/heads/master@{#372676}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 030a0b11cd3b22b114fbbb96979995c8e7ad32a9