Remove find-in-page methods from WebRemoteFrame.
This patch moves all find-in-page methods out of WebRemoteFrame (and out of the generic WebFrame), and into WebLocalFrame specifically. This was done because none of these methods should ever be called on a WebFrame that COULD be remote. They should only ever be called on local frames.

BUG=457440

Review URL: https://codereview.chromium.org/1647623002

Cr-Original-Commit-Position: refs/heads/master@{#372933}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8a2dc803cd6c71f8a5c8e48236335e39ecb5bbf0