Record scrolling layer in ElementLayers whenever possible.
Because Scroll Customization on CompositorWorker needs to be aware of
the main and scrolling layers of the customized element, we need to be
able to get either of them based on the element id.

This is also a bit simpler than the previous solution.

BUG=505851
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1646543003

Cr-Original-Commit-Position: refs/heads/master@{#372940}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 15f972b06df0017e7a873027e395f1a63f02b09c