Add a close() method to ImageBitmap
This is not in the ImageBitmap spec yet, but it is part of the OffScreenCanvas
proposal. In the proposal, the words about this method are:
// It's crucial that there be a way to explicitly dispose of ImageBitmaps
// since they refer to potentially large graphics resources.

So this method is very similar to ImageData::dispose(), which explicitly
dispose any resources that ImageBitmap is associated with. At this moement,
ImageBitmap holds a RefPtr to a StaticBitmapImage, so the close() method
clears the RefPtr.

Since this method is not in the ImageBitmap spec yet, we should hide
it behind a RuntimeEnabled flag.

A layout test is also included in this CL. In particular, the width and
height of the ImageBitmap should be both 0 after calling the close() method.
The layout test also includes cases that calls close() to a already closed
ImageBitmap, or calls close() to a neutered ImageBitmap. In both cases,
the close() operation should be noop.

Now that we have three operations for ImageBitmap: structured clone (SK),
transfering (T) and close (C), we should test the behavior of combining them.
In my previous CLs: https://codereview.chromium.org/1647083003/,
https://codereview.chromium.org/1652553003/,
the cases of calling SK--T and T--SK have been tested. In this CL, the
layout test contains cases for calling SK--C, T--C, C--SK and C--T.

BUG=580649

Review URL: https://codereview.chromium.org/1631763004

Cr-Original-Commit-Position: refs/heads/master@{#373050}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 14dc07c20ddbd9b7324e5a34f35adc3ba917c938