Introduce ScrollAnchor class.
This class adjusts a scroll position to compensate for movement of a visible
node as a result of layout.  See bit.ly/scroll-anchoring for more information.

This patch is just the skeleton; future patches will hook it up to the layout
code and fill out the implementation.

BUG=558575

Review URL: https://codereview.chromium.org/1637483003

Cr-Original-Commit-Position: refs/heads/master@{#373157}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4a58ff9c9c343c2110fa25f7d215f996da772bf6