Re-land: Fix race in UseNewEDK().
The difference with this re-land is that it changes LazyInstance to be leaky to avoid a recursive lock acquisition in AtExitManager. It also doesn't make sense to destroy that instance since it's a cache of a command line flag.

BUG=539315

Review URL: https://codereview.chromium.org/1646223002

Cr-Commit-Position: refs/heads/master@{#372501}

Review URL: https://codereview.chromium.org/1663693002

Cr-Original-Commit-Position: refs/heads/master@{#373176}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6be610508ebbc1787e959e741861f6eecff02c8d