Cache add/addAll should reject if any response is not ok
The Service Worker spec was recently changed [1] to ensure that failed
and/or opaque responses are not cached. If any Response seen by add()
or addAll() is not 'ok' (see Fetch [2]) then the operation should fail
with nothing modified in the cache. Implement that and modify tests.

[1] Spec: https://github.com/slightlyoff/ServiceWorker/issues/823
[2] Fetch: https://fetch.spec.whatwg.org

BUG=583407
R=jkarlin@chromium.org

Review URL: https://codereview.chromium.org/1664473002

Cr-Original-Commit-Position: refs/heads/master@{#373259}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d8a95558a04b5734bc5568546097799d942aaec5