Implementing option imageOrientation for createImageBitmap(ImageData)
This CL implements the ImageBitmap option imageOrientation for the method
createImageBitmap(ImageData) only. For createImageBitmap from other sources,
there will be separate CLs to handle them.

The whatwg wiki for the ImageBitmap options is here:
https://wiki.whatwg.org/wiki/ImageBitmap_Options
In the case of ImageData, the option none/topLeft should behave the same,
which is "Do not change orientation". This is because ImageData does
not contain any image media meta data. Similarly, For the option of
flipY/topLeft, the behavior should be flip the image vertically.

This CL also includes a layout test, which re-uses the existing test with
the additional parameter of "flipY", and currently the layout test is for
the ImageData as source only.

BUG=249382

Review URL: https://codereview.chromium.org/1631733003

Cr-Original-Commit-Position: refs/heads/master@{#373274}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 542c67487c8bac201e42f607f6f3727382831c97