Fix onerror event for link preload raw resources
Link preload resources that were fetched using RawResource (so media,
track and resources with no `as` value), failed to trigger the onerror
event as a result of HTTP errors.

This CL fixes that. It also renames the previous avoidBlockingLoadEvent
flag, to make it clearer that the resource is a link preload one.

BUG=

Review URL: https://codereview.chromium.org/1653193002

Cr-Original-Commit-Position: refs/heads/master@{#373316}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a416c62387b5ac9a3e91b9003f9969585dc51792