Make decoding of createImageBitmap(Blob) work on a background thread
At this moment, the decoding step of this function always happens on the
same thread as the thread that calls this function. For example, if
createImageBitmap(Blob) is performed on the main thread, then decoding
happens on the main thread which could jank the main thread for large
images. This CL intends to move the decoding part to a background thread.

For this particular change, as long as all the current layout tests
(particularly the ones of createImageBitmap-blob) pass, it should not
cause any problem.

BUG=580202

Review URL: https://codereview.chromium.org/1644433003

Cr-Original-Commit-Position: refs/heads/master@{#373513}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68bba9563388e23e91e808c91d5ad60f58731fdb