Fix viewport resize notification for images with sizes and srcset
Unregister viewport listener if new image isn't viewport dependant. This was
wasteful as each resize would cause the image to try selecting a new image
source even though the image isn't viewport dependent.

Move the viewport listener registration to happen in
setBestFitURLAndDPRFromImageCandidate since this method is called anytime we
change image sources. When an <img> with a viewport descriptor in a <picture>
element is added, it's initialized through HTMLImageElement::insertedInto
which would fail to register the viewport listener if one hadn't already been
created - such as the case for page load.

Move notifyResizeForViewportUnits back into performPreLayoutTasks. I'm moving
this back to where I found it as it's the cause of the attached bug. The layout
size won't change when the browser zoom level is changed but we need to recalc
vh/vw units at that time. This was previously moved here by me in Blink
r190488.

BUG=573187

Review URL: https://codereview.chromium.org/1650083003

Cr-Original-Commit-Position: refs/heads/master@{#373523}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f2d6243b2c6092eb70ccdc6a98e13ea04992ecf0