Code cleanup: Remove InterpolationType from ConversionChecker constructor
There are no behavioural changes in this patch.

This change removes the requirement for ConversionChecker objects to be
constructed with their associated InterpolationType and defers the
association to InvalidatableInterpolation.

This reduces boilerplate in an already boilerplate heavy class.

Additionally the trace method has been removed from ConversionChecker
as it is not garbage collected and will never be called.

Review URL: https://codereview.chromium.org/1609693002

Cr-Original-Commit-Position: refs/heads/master@{#373675}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b72dcae8a0f927d181f48507737c6ce5e33e4a0a