Fix subpixel text anti-aliasing with display list 2D canvas
This CL triggers a display list fallback when the canvas is opaque
(which triggers subpixel AA) and text is drawn to the canvas. This
is a short term fix. The long term fix is to make display lists support
subpixel AA, so that a fallback is unnecessary. See: skbug.com/4897

BUG=583809

Review URL: https://codereview.chromium.org/1670723002

Cr-Original-Commit-Position: refs/heads/master@{#373692}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b67e74eff228013993614bccfd9ee83a65cd494