Don't force display:list-item to display:block in quirks mode when float is set to a value
Currently in quirks mode, and element with float set and display set to
list-item is forced to be display:block. This current action is not
supported in quirks spec and is inconsistent with FF/Edge. This patch fixes
that behavior and allows these elements to retain their display value
even in quirks mode.

BUG=580077

Review URL: https://codereview.chromium.org/1620573002

Cr-Original-Commit-Position: refs/heads/master@{#374115}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 06ae6f0489cf8efa87e7190ab90940b62a998c8b