Implement ImageBitmap options premultiplyAlpha
In this CL, we implement the option of premultiplyAlpha for ImageBitmap.
At this moment, it works for createImageBitmap from HTMLImageElement,
HTMLCanvasElement and Blob. The cases of from ImageBitmap, HTMLVideoElemtn
and ImageData will be solved in future CL(s).

This CL also includes modifies a current layout tests, which tests
the combination of ImageBitmap options "flipY" and premultiplyAlpha

BUG=249382

Review URL: https://codereview.chromium.org/1609763002

Cr-Original-Commit-Position: refs/heads/master@{#374132}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dfbf8dd24ba36edab0e567c3f98f4ed6ad8dd1e2