Make Resource RefCountedWillBeGarbageCollectedFinalized, attempt #2
Resource's lifetime is currently managed by an unholy mix of
factors that would ideally be represented in a more canonical
way. Historically, some of this was necessary, and it was
glossed over because of ResourcePtr which, in addition to
keeping the Resource alive, performed a key part of
revalidation handling in the case of a 304. Now that
ResourcePtr's revalidation behavior has been removed,
standardize Resource's lifetime management.

BUG=571190

Review URL: https://codereview.chromium.org/1667843003

Cr-Original-Commit-Position: refs/heads/master@{#374220}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 486e98700f552d7a5c4fbc1b3a9b5e9f2a6cf934