WTF::bind: Remove the need of hand-rolled code for bound arguments.
This patch adds a generalized version of PartBoundFunctionImpl for
WTF::bind(), and removes the hand-rolled code that contains the
repetition for cases with zero to seven bound arguments.

The key trick is to use std::tuple<T...> to pass the types of bound
parameters.

This patch adds a new dependency to base/tuple.h for a simple utility
called base::IndexSequence defined there. The bind code lives in WTF, so
this should be fine.

BUG=565765

Review URL: https://codereview.chromium.org/1679773002

Cr-Original-Commit-Position: refs/heads/master@{#374319}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2a02e48ea74b0608a2eb43aa935a3cd7dbaea8f2