Implement ImageBitmap options premultiplyAlpha for HTMLVideoElement
In this CL, we implement the premultiplyAlpha option for ImageBitmap in
the case of HTMLVideoElement.

This CL modifies the existing layout test to test all the cases of
cobmining flipY and premultiplyAlpha. Several cases are commentted out
in the layout test because skia doesn't yet support draw an umpremul
SkImage.

BUG=249382

Review URL: https://codereview.chromium.org/1677383002

Cr-Original-Commit-Position: refs/heads/master@{#374511}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c85a909181ade63f293b499b1a290ab11f6e59f9