[Experimental Framework] Add version field to trial tokens - Reland
This is a reland of https://codereview.chromium.org/1676513002/.

Add a version number to the token format, to allow for incompatible
changes to the format over time. The version number is a strictly
increasing integer, not linked to Chrome versions.

The version number is not included in the signed data, or used in the
signature validation. The assumption is that it's not feasible for an
attacker to obtain a token for version "x", then modify the data
and/or signature, set the version to "y" and have that be recognized
as a valid token.  Specifically, the main motivation for the version
number is to allow changes to the signature algorithm. When bumping
the version number, the expected practice would be to change the
signing keys as well to mitigate cross-version attacks.

Only one version is currently supported: "1". Other than parsing,
validating and storing the version, this CL does not introduce support
for having multiple versions of the parsing or validation algorithms.
Multiple version support can be added if/when changes to the token format
are required.

BUG=570684

Review URL: https://codereview.chromium.org/1684133002

Cr-Original-Commit-Position: refs/heads/master@{#375130}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 29fd1e1f43a13a74e7836722bf9dea2b0ec44729